Exploring many paths, finding one that sticks
Of all the paths tested for bundling/hydration, the one that looks like a winner is a single entry at main.ts, where import paths for each component must be explicitly and dynamically imported, to produce separate modules. This method does not require any html entrypoint (which would have been hacky anyway), but requires us to know ahead of time what each relative path will be for each component tree we want. The probable path to automation is a custom vite plugin which will do a bit of code-mod on main.ts to set up the import paths, or potentially even write the main entry file entirely.

Dev mode should work by importing literal .svelte components from src directory, but needs "main.ts" to be added as this is where the "hydrate" function is defined (but we will figure this part out later).