Set up standalone vite plugin, mostly complete.
The issue here is that dev/serve and build have to be handled separately, so the user still has a direct dependency on svelte plugin in their own vite config, and the config is not a straightforward as desired. This does work though, for dev and prod 'archipelago' workflows. This can be tested by running dev and visiting /dev.html, and build/preview and visiting /prod.html